Make it easier to test tasks
This allows the task to be tested whilst ensuring the before and after
task functions are run. This makes the task easier to test as these
setup and teardown functions likely enable aspects that the task needs
to run.

This is similar to a Quart route and the need to use a test client to
test them to ensure the before/after request functions are called.